requirements.txt
httpx
torch
pydub
python-dotenv
customtkinter
tkinterdnd2